refactor(envelope): split artifact owner
