+++ crates/tokmd-envelope/src/lib.rs
@@ -17,8 +17,11 @@
 //! * tokmd-specific analysis types (use tokmd-analysis-types)
 //! * I/O operations or business logic
 
+mod artifact;
 pub mod ffi;
 pub mod findings;
+
+pub use artifact::Artifact;
 
 use serde::{Deserialize, Serialize};
 
@@ -271,34 +274,6 @@ pub struct GateItem {
     /// Path to the source artifact (if applicable).
     #[serde(skip_serializing_if = "Option::is_none")]
     pub artifact_path: Option<String>,
-}
-
-/// Artifact reference in the sensor report.
-///
-/// # Examples
-///
-/// ```
-/// use tokmd_envelope::Artifact;
-///
-/// let art = Artifact::receipt("output/receipt.json")
-///     .with_id("analysis")
-///     .with_mime("application/json");
-/// assert_eq!(art.artifact_type, "receipt");
-/// assert_eq!(art.id.as_deref(), Some("analysis"));
-/// ```
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct Artifact {
-    /// Artifact identifier (e.g., "analysis", "handoff").
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub id: Option<String>,
-    /// Artifact type (e.g., "comment", "receipt", "badge").
-    #[serde(rename = "type")]
-    pub artifact_type: String,
-    /// Path to the artifact file.
-    pub path: String,
-    /// MIME type (e.g., "application/json").
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub mime: Option<String>,
 }
 
 /// Status of a capability for "No Green By Omission".
@@ -631,45 +606,6 @@ impl GateItem {
     }
 }
 
-impl Artifact {
-    /// Create a new artifact reference.
-    pub fn new(artifact_type: impl Into<String>, path: impl Into<String>) -> Self {
-        Self {
-            id: None,
-            artifact_type: artifact_type.into(),
-            path: path.into(),
-            mime: None,
-        }
-    }
-
-    /// Create a comment artifact.
-    pub fn comment(path: impl Into<String>) -> Self {
-        Self::new("comment", path)
-    }
-
-    /// Create a receipt artifact.
-    pub fn receipt(path: impl Into<String>) -> Self {
-        Self::new("receipt", path)
-    }
-
-    /// Create a badge artifact.
-    pub fn badge(path: impl Into<String>) -> Self {
-        Self::new("badge", path)
-    }
-
-    /// Set the artifact ID. Builder pattern.
-    pub fn with_id(mut self, id: impl Into<String>) -> Self {
-        self.id = Some(id.into());
-        self
-    }
-
-    /// Set the MIME type. Builder pattern.
-    pub fn with_mime(mut self, mime: impl Into<String>) -> Self {
-        self.mime = Some(mime.into());
-        self
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -967,24 +903,5 @@ mod tests {
         assert_eq!(gate.reason.as_deref(), Some("Below threshold"));
         assert_eq!(gate.source.as_deref(), Some("ci_artifact"));
         assert_eq!(gate.artifact_path.as_deref(), Some("coverage/lcov.info"));
-    }
-
-    #[test]
-    fn artifact_builders_cover_variants() {
-        let custom = Artifact::new("custom", "out/custom.json");
-        assert_eq!(custom.artifact_type, "custom");
-        assert_eq!(custom.path, "out/custom.json");
-
-        let comment = Artifact::comment("out/comment.md");
-        assert_eq!(comment.artifact_type, "comment");
-        assert_eq!(comment.path, "out/comment.md");
-
-        let receipt = Artifact::receipt("out/receipt.json");
-        assert_eq!(receipt.artifact_type, "receipt");
-        assert_eq!(receipt.path, "out/receipt.json");
-
-        let badge = Artifact::badge("out/badge.svg");
-        assert_eq!(badge.artifact_type, "badge");
-        assert_eq!(badge.path, "out/badge.svg");
     }
 }
